transcendence/requirements/local.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
coverage==7.3.3
cryptography==41.0.7
daphne==4.0.0
Django==4.2.8
django-formtools==2.5.1
django-otp==1.3.0
django-phonenumber-field==7.2.0
django-two-factor-auth==1.15.5
factory-boy==3.3.0
Faker==21.0.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
mock==5.1.0
msgpack==1.0.7
phonenumberslite==8.13.27
psycopg==3.1.14
psycopg-binary==3.1.14
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.3.0
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.0
qrcode==7.4.2
redis==5.0.1
requests==2.31.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
urllib3==2.1.0
zope.interface==6.1
